src/requirements.txt
protobuf==3.20.3
pyyaml
mlflow
pandas
numpy
scikit-learn
h3
joblib
